sytanx: Rewrite the test section matching
This is a complete rewrite, rather than aiming to be as simple as
possible, it aims to be as correct as possible. The regular expressions
are quite complex, but it is able to accomplish much more than the
previous attempt did.

TODO: This is only about 50% complete at the moment.